`--argc-build` add `[OUTPATH]` parameter
